Overall Optimization. Finished Series By Votes.
Added Test module. Added utility function in cota-db findById(id, array). Moved Group ID generation to a separate function in cota-db generateGroupId(). Refactored most for loops to maps. Added an additional verification to the min and max params in GroupSeriesByVotes so we don't allow anything that isn't a number as  value. Refactored every "serie" word to "series" as "series" is both singular and plural.